requirements.txt
absl-py==0.9.0
astor==0.8.1
astroid==2.3.3
atari-py==0.2.6
cloudpickle==1.3.0
cycler==0.10.0
future==0.18.2
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.2
gym==0.17.1
h5py==2.10.0
isort==4.3.21
joblib==0.14.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
Markdown==3.2.1
matplotlib==3.2.0
mccabe==0.6.1
numpy==1.18.1
opencv-python==4.2.0.32
pandas==1.0.1
Pillow==7.0.0
protobuf==3.11.3
pybullet==2.6.6
pyglet==1.5.0
pyinter==0.2.0
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
scipy==1.4.1
six==1.14.0
stable-baselines==2.10.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
tqdm==4.43.0
typed-ast==1.4.1
Werkzeug==1.0.0
wrapt==1.12.1
